pyproject.toml
[tool.poetry]
name = "mitsuakybot"
version = "0.1.0"
description = "Personal Discord bot for servers I like"
authors = ["Pármenas Haniel <parmenashp@gmail.com>"]
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git"}
asyncpg = "^0.26.0"
loguru = "^0.6.0"
tomli = "^2.0.1"
jishaku = "^2.5.0"
prisma = "^0.12.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[tool.black]
line-length = 119

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
